+++ .example.env
@@ -1,0 +1,1 @@
+BLADE_APP_TOKEN="000"
