+++ components/link.client.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import type { ReactNode } from 'react';
 
 const twitterURL = 'https://twitter.com/notquiteleo';
@@ -33,6 +31,7 @@ const Link = ({ children, href, toggleEmail }: LinkProps) => (
   <a
     href={href || twitterURL}
     target="_blank"
+    rel="noreferrer"
     onMouseEnter={toggleEmail && onMouseEnter}
     onMouseLeave={toggleEmail && onMouseLeave}
     className="text-base no-underline text-black p-2.5 block md:rounded-md md:py-1 md:px-2.5 md:mx-1.5 md:transition md:duration-200 md:hover:bg-neutral-200/60"
