+++ package.json
@@ -1,10 +1,12 @@
 {
   "name": "leo",
-  "version": "300.0.0",
+  "version": "400.0.0",
+  "type": "module",
   "license": "MIT",
   "scripts": {
-    "dev": "next",
-    "build": "next build",
+    "dev": "blade",
+    "build": "blade build",
+    "serve": "blade serve",
     "prepare": "husky"
   },
   "repository": "leo/site",
@@ -12,21 +14,19 @@
     "*.{js,jsx,ts,tsx,css,md}": "prettier --write --ignore-path .gitignore"
   },
   "dependencies": {
+    "@ronin/blade": "0.2.1",
     "clsx": "2.1.1",
-    "next": "14.2.4",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
-    "ronin": "6.4.1"
+    "react": "0.0.0-experimental-df12d7eac-20230510",
+    "react-dom": "0.0.0-experimental-df12d7eac-20230510",
+    "ronin": "6.4.20",
+    "tailwindcss": "3.4.7"
   },
   "devDependencies": {
-    "@types/node": "20.14.2",
-    "@types/react": "18.3.3",
-    "autoprefixer": "10.4.19",
+    "@types/react": "19.1.2",
+    "@types/react-dom": "19.1.2",
     "husky": "9.0.11",
-    "lint-staged": "15.2.7",
-    "postcss": "8.4.38",
-    "prettier": "3.3.2",
-    "tailwindcss": "3.4.4",
-    "typescript": "5.8.2"
+    "lint-staged": "15.5.1",
+    "prettier": "3.5.3",
+    "typescript": "5.8.3"
   }
 }
