+++ tsconfig.json
@@ -20,13 +20,8 @@
       "@pages/*": ["app/*"],
       "@styles/*": ["styles/*"],
       "@schema": ["schema/index"]
-    },
-    "plugins": [
-      {
-        "name": "next"
-      }
-    ]
+    }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", ".ronin/*.d.ts"],
+  "include": ["**/*.ts", "**/*.tsx", ".ronin/*.d.ts"],
   "exclude": ["node_modules"]
 }
